+++ src/core/Pile.gd
@@ -40,6 +40,15 @@ func _process(_delta) -> void:
 			obj.queue_free()
 	# We make sure to adjust our popup if cards were removed from it while it's open
 	$ViewPopup.set_as_minsize()
+	if get_card_count() > 0 and cfc.focus_style:
+		var top_card = get_top_card()
+		if cfc.NMAP.board.mouse_pointer in get_overlapping_areas()\
+				and not cfc.card_drag_ongoing:
+			if (top_card.is_faceup or top_card.is_viewed)\
+					and top_card.state == Card.CardState.IN_PILE:
+				top_card.state = Card.CardState.VIEWED_IN_PILE
+		elif top_card.state == Card.CardState.VIEWED_IN_PILE:
+			top_card.state = Card.CardState.IN_PILE
 
 
 # Populates the popup view window with all the cards in the deck